+++ src/Commands/DebugCheats.cs
@@ -31,16 +31,18 @@ namespace PressureDrop.Commands
                 case "?":
                 case "h":
                 case "help":
-                    ChatCommander.Output($"Commands provided by <style=cSub>{Plugin.GUID}</style>:");
+                    ChatCommander.Output($"<style=cWorldEvent>{Plugin.GUID}</style> debug cheats:");
                     ChatCommander.Output($"  <style=cSub>{ChatCommander.commandPrefix}aq</style>: changes the stage to Abandoned Aqueduct.");
                     ChatCommander.Output($"  <style=cSub>{ChatCommander.commandPrefix}g</style>: gives the user items for testing pressure plate changes.");
                     ChatCommander.Output($"  <style=cSub>{ChatCommander.commandPrefix}p</style>: drops items for testing recyclability config.");
-                    ChatCommander.Output($"  <style=cSub>{ChatCommander.commandPrefix}ne</style>: disables further enemy spawns.");
+                    ChatCommander.Output($"  <style=cSub>{ChatCommander.commandPrefix}ne</style>: toggles enemy spawns.");
                     break;
             }
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("", "Publicizer001")]
+
+
+
         private static void ForceAqueduct(string[] args)
         {
             if (args.Length > 1) {
@@ -48,47 +50,30 @@ namespace PressureDrop.Commands
                 return;
             }
 
-            SceneDef scene = SceneCatalog.FindSceneDef("goolake");
             // Run.instance.AdvanceStage(scene);
+#pragma warning disable Publicizer001 // Accessing a member that was not originally public
             Run.instance.GenerateStageRNG();
-            UnityEngine.Networking.NetworkManager.singleton.ServerChangeScene(scene.cachedName);
+#pragma warning restore Publicizer001 // Accessing a member that was not originally public
+            UnityEngine.Networking.NetworkManager.singleton.ServerChangeScene("goolake");
 
             Chat.SendBroadcastChat(new Chat.SimpleChatMessage { baseToken = "<style=cWorldEvent>Sending you to the Origin of Tar <sprite name=\"Skull\" tint=1></style>" });
         }
 
         private static void GiveCommand(NetworkUser user, string[] args)
         {
-            string invalid = "expects one argument from the following options [<style=cSub>t</style>, <style=cSub>s</style>, <style=cSub>g</style>, <style=cSub>e</style>]";
+            string invalid = "{ <style=cSub>e</style> | <style=cSub>t</style> | <style=cSub>s</style> | <style=cSub>g</style> } [<style=cSub><count></style>]";
 
             const int expectedArgs = 2;
             if (args.Length >= expectedArgs) {
-                ItemDef item;
-                int count;
+                ParseGiveCommandItem(args[1].ToLowerInvariant(), out ItemDef item, out int count, out EquipmentDef equipment);
+                if (item == null) item = ItemCatalog.GetItemDef(ItemCatalog.FindItemIndex(args[1]));
 
-                switch (args[1].ToLowerInvariant()) {
-                    default:
-                        ChatCommander.OutputFail(args[0], invalid);
-                        return;
-                    case "e":
-                        {
-                            EquipmentDef e = RoR2Content.Equipment.Scanner;
-                            user.master.inventory.GiveEquipmentString(e.name);
-                            ChatCommander.Output($"Gave {user.userName} {ChatCommander.GetColoredPickupLanguageString(e.equipmentIndex)}");
-                            return;
-                        } 
-                    case "t":
-                        item = RoR2Content.Items.Bear;
-                        count = 70;
-                        break;
-                    case "s":
-                        item = RoR2Content.Items.SprintBonus;
-                        count = 5;
-                        break;
-                    case "g":
-                        item = RoR2Content.Items.AutoCastEquipment;
-                        count = 7;
-                        break;
+                if (equipment != null) {
+                    user.master.inventory.GiveEquipmentString(equipment.name);
+                    ChatCommander.Output($"Gave {user.userName} {ChatCommander.GetColoredPickupLanguageString(equipment.equipmentIndex)}");
+                    return;
                 }
+                else if (item == null) { ChatCommander.OutputFail(args[0], invalid); return; }
 
                 if (args.Length > expectedArgs && int.TryParse(args[2], out int num)) count = num;
                 if (count > 1000) count = 1000; // Let's not get too excessive
@@ -99,9 +84,38 @@ namespace PressureDrop.Commands
             else ChatCommander.OutputFail(args[0], invalid);
         }
 
+        private static void ParseGiveCommandItem(string arg, out ItemDef item, out int count, out EquipmentDef equipment)
+        {
+            item = null;
+            equipment = null;
+            count = 1;
+
+            switch (arg) {
+                default: break;
+                case "e":
+                    equipment = RoR2Content.Equipment.Scanner;
+                    break;
+                case "t":
+                    item = RoR2Content.Items.Bear;
+                    count = 70;
+                    break;
+                case "s":
+                    item = RoR2Content.Items.SprintBonus;
+                    count = 5;
+                    break;
+                case "g":
+                    item = RoR2Content.Items.AutoCastEquipment;
+                    count = 7;
+                    break;
+            }
+        }
+
+
+
+
         private static void PickupCommand(NetworkUser user, string[] args)
         {
-            string invalid = "expects one argument from the following options [<style=cSub>i</style>, <style=cSub>e</style>, <style=cSub>s</style>, <style=cSub>v</style>, <style=cSub>h</style>]";
+            string invalid = "{ <style=cSub>t</style> | <style=cSub>v</style> | <style=cSub>s</style> }";
 
             const int expectedArgs = 2;
             if (args.Length == expectedArgs) {
@@ -111,33 +125,18 @@ namespace PressureDrop.Commands
                     default:
                         ChatCommander.OutputFail(args[0], invalid);
                         return;
-                    case "i":
+                    case "t":
                         PressureDrop.Drop.DropStyleChest(target, [
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.SprintBonus.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.EquipmentMagazine.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.ExtraLife.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.SprintWisp.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.AutoCastEquipment.itemIndex),
-                            PickupCatalog.FindPickupIndex(RoR2Content.Items.Pearl.itemIndex)
-                        ], 3.4f, 14f);
-                        ChatCommander.Output($"Dropped item pickups @ {user.userName}.");
-                        break;
-                    case "e":
-                        PressureDrop.Drop.DropStyleChest(target, [
+                            PickupCatalog.FindPickupIndex(RoR2Content.Items.Pearl.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Equipment.Recycle.equipmentIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Equipment.Tonic.equipmentIndex),
-                            PickupCatalog.FindPickupIndex(RoR2Content.Equipment.AffixRed.equipmentIndex)
                         ], 3.4f, 14f);
-                        ChatCommander.Output($"Dropped equipment pickups @ {user.userName}.");
-                        break;
-                    case "s":
-                        PressureDrop.Drop.DropStyleChest(target, [
-                            PickupCatalog.FindPickupIndex(RoR2Content.Items.ScrapWhite.itemIndex),
-                            PickupCatalog.FindPickupIndex(RoR2Content.Items.ScrapGreen.itemIndex),
-                            PickupCatalog.FindPickupIndex(RoR2Content.Items.ScrapRed.itemIndex),
-                            PickupCatalog.FindPickupIndex(RoR2Content.Items.ScrapYellow.itemIndex)
-                        ], 3.4f, 14f);
-                        ChatCommander.Output($"Dropped scrap pickups @ {user.userName}.");
+                        ChatCommander.Output($"Dropped test pickups @ {user.userName}.");
                         break;
                     case "v":
                         PressureDrop.Drop.DropStyleChest(target, [
@@ -148,7 +147,7 @@ namespace PressureDrop.Commands
                         ], 3.4f, 14f);
                         ChatCommander.Output($"Dropped void pickups @ {user.userName}.");
                         break;
-                    case "h":
+                    case "s":
                         PressureDrop.Drop.DropStyleChest(target, [
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.ExtraLifeConsumed.itemIndex),
                             PickupCatalog.FindPickupIndex(RoR2Content.Items.TonicAffliction.itemIndex),
@@ -163,6 +162,9 @@ namespace PressureDrop.Commands
             else ChatCommander.OutputFail(args[0], invalid);
         }
 
+
+
+
         private static void DisableEnemySpawns(string[] args)
         {
             if (args.Length > 1) {
@@ -171,9 +173,7 @@ namespace PressureDrop.Commands
             }
 
             bool wasDisabled = CombatDirector.cvDirectorCombatDisable.GetString() != "0";
-
             CombatDirector.cvDirectorCombatDisable.SetBool(!wasDisabled);
-
             if (wasDisabled) ChatCommander.Output("Enemy spawns enabled.");
             else ChatCommander.Output("Enemy spawns disabled.");
         }